feat: improve dogdrip collector and introduce log-scale buzz score system
- Switch target board from /boomupbest to main /dogdrip board

- Parse comments and upvotes from HTML lists

- Change tag count fields from int to float64 for log scaling

- Calculate log-scale buzz scores on Go backend using comments/upvotes

- Fix integration test compiler errors and race conditions

+++ server/domain/communitytrend/worksheet.go
@@ -21,8 +21,8 @@ type Worksheet struct {
 
 // TagCount is one confirmed (tag, count) pair for a community-day.
 type TagCount struct {
-	TagID int `json:"tag_id"`
-	Count int `json:"count"`
+	TagID int     `json:"tag_id"`
+	Count float64 `json:"count"`
 }
 
 // Confirmation is the atomic payload written when a worksheet is confirmed.
